requirements.txt
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
beautifulsoup4==4.9.3
blis==0.7.4
catalogue==1.0.0
certifi==2020.12.5
chardet==4.0.0
cymem==2.0.5
distlib==0.3.1
Django==3.1.4
django-bootstrap4==2.3.1
django-filter==2.4.0
djangorestframework==3.12.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz
filelock==3.0.12
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
Markdown==3.3.3
mccabe==0.6.1
murmurhash==1.0.5
numpy==1.19.4
packaging==20.8
Pillow==8.0.1
plac==1.1.3
pluggy==0.13.1
preshed==3.0.5
py==1.10.0
pylint==2.6.0
pyparsing==2.4.7
pytest==6.2.1
pytz==2020.5
requests==2.25.1
setuptools==44.0.0
six==1.15.0
soupsieve==2.1
spacy==2.3.5
sqlparse==0.4.1
srsly==1.0.5
thinc==7.4.5
toml==0.10.2
tox==3.20.1
tqdm==4.55.0
urllib3==1.26.2
virtualenv==20.2.2
wasabi==0.8.0
wheel==0.36.2
wikipedia==1.4.0
wrapt==1.12.1
